+++ src/pages/play-video/VideoInfo/video-info-styles.css
@@ -42,10 +42,12 @@
 .action-buttons {
   display: flex;
   gap: 3rem;
+  padding: 0.5rem;
+  margin-bottom: 0.5rem;
 }
 
 @media (max-width: 640px) {
   .action-buttons {
-    gap: 1rem;
+    justify-content: space-between;
   }
 }
