+++ src/pages/play-video/video-player-page-styles.css
@@ -1,8 +1,14 @@
 .video-player {
   width: 80%;
   margin: auto;
-  margin-top: 1rem;
+}
+
+.react-player {
   height: 90vh;
+}
+
+.react-player-mobile {
+  display: none;
 }
 
 .input-text {
@@ -25,6 +31,15 @@
 @media (max-width: 640px) {
   .video-player {
     width: 95%;
-    height: auto;
+  }
+
+  .react-player-desktop {
+    display: none;
+  }
+
+  .react-player {
+    height: 250px;
+    width: 95%;
+    margin: auto;
   }
 }
